Extract year from strain
Some sequences have incorrect date information in the Pathoplexus
metadata which becomes apparent when viewed on a time tree. For these
sequences, the strain name has a year-only date which matches related
sequences on the time tree, so parse that out and replace the other date
information.

Brief preview:

    'PP_000M7HH': date '2011-07-24' → '1995-XX-XX'
    'PP_000MBZF': date 'XXXX-XX-XX' → '1995-XX-XX'
    'PP_000MC0D': date 'XXXX-XX-XX' → '1995-XX-XX'

See logs/extract_from_strain.txt for a full list.