Restore phylogenetic workflow
The workflow was broken by the previous commit that shuffled things
around to conform to pathogen repo guide. This restores the workflow to
a working state.

Note that, while the workflow aims to replace what's currently on
nextstrain.org/ebola, as of now it is nowhere close to replicating that
dataset given the changes to the underlying data source (now
Pathoplexus). The main motivation of having this in a working state is
to eventually make another dataset. This will come in another PR.