Fix broken URL
The URL had previously worked with 'refs/heads/master'. I replaced
'master' with a commit hash in an attempt to create a permalink, but did
not notice that it was invalid since curl was not exiting with a
failure. Now that it does, the mistake is obvious.

New URL retrieved from the GitHub website.