ingest: Only download latest version of sequences
I noticed a large number of sequences without matching metadata and realized 
the workflow was download all versions of sequences but only the latest version 
of metadata. I can't find any documentation on the `versionStatus` query param,
but local testing show it also works for sequences. This reduced the runtime 
from 83s to 24s and reduced file size from 200M to 63M.