CI: run on push to new default branch `main`
The default branch for this repo was renamed from `master` to `main` as 
noted in https://github.com/nextstrain/ebola/pull/14.

Unfortunately there's no variable for the default branch that can be used in 
the `on.push` block, so this has to be manually maintained. We could add a 
conditional for `jobs.ci` to only run on the default branch, but this would 
cause the CI workflow to trigger on every push and just skip the `ci` job. 
I think it's cleaner to just manually maintain the default branch name since 
it's unlikely that we'll change it again.