Update pre-commit config
to remove the end-of-line & newline at end-of-file checks. We can
restore these if we run into issues in the future (e.g. running `cat`
on files which don't finish with a newline).

Also adds ruff configuration and disables the rule F541
(f-string without any placeholders) which is unnecessary in my eyes
because it's not an error