Don't override the workdir
Keep the working directory from where we invoke snakemake from (or which
we set via `--directory`). Note that I expect the working directory
to be `phylogenetic` when run within the ebola repo or an external
analysis directory (not yet supported).

(This commit breaks the workflow but subsequent commits will fix it)