Move config.files to rule-specific values
We've used 'files' ~forever but I find rule-specific config paths much
more interpretable.

This has the slight oddity that the alignment reference is used for
translation as well, which is non-obvious from the config structure. I
explored duplicating this (via YAML anchors) into
`config.translate.reference` but this exposed the footgun of using
different references for different steps -- see
<https://github.com/nextstrain/ebola/pull/19#discussion_r2395973438>