Update filtering logic
to use the outbreak metadata (assigned via Nextclade in ingest), which
follows the approach we use in the all-outbreaks dataset where we
use this data to subsample on.

I played around with `--exclude-all` and `--include-where` arguments
but found them too limited as I also wanted to exclude lab hosts,
whereas the `--query` argument is both powerful enough to do it all
itself if needed and also doesn't need the `--exclude-all` argument.

Finally the verbosity of the `rule filter` is striking. Ideally we want
the config YAML to be able to express anything that `augur filter` can
handle, but currently we need a lot of conditional logic in the
snakemake rule to allow this. I believe the `augur subsample` config
YAML will give us this.