PROTOTYPE config pre-parsing
Our current approach of connecting config YAML keys/values to a command
line invocation requires a large amount of in-the-middle code, notably
the params block in the snakemake rule. A good example of this is in
one of the grandparent commits <https://github.com/nextstrain/ebola/pull/19/commits/37aced59c0a54556dff571098fda5c4eb3889bf8>

This in-the-middle-code is both painful to write and inherently per-repo.

Here I explore a different approach where we up-front parse the config
YAML into per-rule command-line arguments, and resolve paths as needed.
This code, and the YAML key name to arg name mappings could be
centralised or read in from a schema (etc), which has the additional
benefit of consistency in YAML key names across repos. Essentially we
shift the cost to a centralised bit of code and make the per-repo 'rule'
block much simpler to write, and more powerful as it has access to a
larger range of config options than we'd typically write. (E.g. no more
commits like the one linked above.) Furthermore nothing here stops a
repo-specific addition to the config, or repo-specific rules /
functionality.

A side-effect of this implementation is that it leads us towards a YAML
structure where each rule has an associated block, which is the dominant
style in many repos (ebola, avian-flu, etc) but we often store keys such
as 'id_column', 'files.x' elsewhere. With YAML anchors I think it's
plausible that we have all config variables for a rule stored under the
rule's name, which is clarifying for the author and should help with our
efforts to document and describe the config YAML.

-------------------------  HOW TO RUN  ------------------------
I've only implemented this prototype for the outbreak-specific phylo
workflow (2 build wildcards) and for rules up to `refine`. Once you've
run the ingest workflow to provision files you can run:

snakemake --cores 1 --snakefile outbreak-specific/Snakefile \
  -pf results/drc-2025/tree.nwk results/west-africa-2014/tree.nwk

To test an external analysis directory you'll first need to
copy the (ingest) sequences.fasta & metadata.tsv into the directory
and create an overlay YAML like so:

inputs:
  - name: mine
    metadata: metadata.tsv
    sequences: sequences.fasta

Then run via the above snakemake command, but updating `--snakefile`
to point to the outbreak-specific/Snakefile in the ebola repo